feat: working review creation

+++ src/main/java/ca/ulaval/glo2003/controllers/requests/ReviewRequest.java
@@ -5,7 +5,7 @@ import ca.ulaval.glo2003.service.dtos.CustomerDTO;
 public record ReviewRequest(
         String date,
         CustomerDTO customer,
-        Integer rating,
+        double rating,
         String comment) {
 
 }
