+++ src/main/java/ca/ulaval/glo2003/domain/review/Review.java
@@ -10,10 +10,10 @@ public class Review {
     private String id;
     private String date;
     private Customer customer;
-    private Integer rating;
+    private double rating;
     private String comment;
 
-    public Review(String restaurantId, String id, String date, Customer customer, Integer rating, String comment) {
+    public Review(String restaurantId, String id, String date, Customer customer, double rating, String comment) {
         this.restaurantId = restaurantId;
         this.id = id;
         this.date = date;
@@ -22,7 +22,7 @@ public class Review {
         this.comment = comment;
     }
 
-    public Review(String restaurantId, String date, Customer customer, Integer rating, String comment) {
+    public Review(String restaurantId, String date, Customer customer, double rating, String comment) {
         this.restaurantId = restaurantId;
         this.id = UUID.randomUUID().toString();
         this.date = date;
@@ -46,7 +46,7 @@ public class Review {
         return customer;
     }
 
-    public Integer getRating() {
+    public double getRating() {
         return rating;
     }
